Flask/requirements.txt
Flask
Flask-SQLAlchemy
mysql-connector-python
requests
python-dotenv
